Ajustes do css
