+++ src/index.html
@@ -4,7 +4,7 @@
     <meta charset="UTF-8" />
     <meta name="viewport" content="width=device-width, initial-scale=1.0" />
     <title>Bug dos Milennialls</title>
-    <link rel="stylesheet" href="./lib/Tela Inicial/telaInicial.css">
+    <link rel="stylesheet" href="./lib/TelaInicial/telaInicial.css">
     <link rel="stylesheet" href="./lib/Cadastro/cadastro.css">
