+++ src/lib/TelaInicial/telaInicial.css
@@ -1,4 +1,4 @@
-@import url('https://fonts.googleapis.com/css2?family=Montserrat:ital,wght@1,600&family=Quicksand:wght@400;600;700&display=swap');
+@import url('https://fonts.googleapis.com/css2?family=Montserrat:ital,wght@0,800;0,900;1,600&family=Quicksand:wght@400;600;700&display=swap');
 
 body {
     background-color: #D9BACA;
@@ -6,44 +6,52 @@ body {
 
 #tituloPagina {
     color: #D2398A;
-    font-family: Montserrat, sans-serif;
-    font-weight: 700;
+    font-family: 'Montserrat', sans-serif;
+    font-weight: 800;
     font-style: normal;
-    filter: blur (2px);
+    filter: blur(1px);
     text-align: center;
-    padding-top: 100px;
-
+    padding-top: 70px;
+    border-bottom: 2px solid #D2398A;
+    text-shadow: 0 0 10px #a75380;
+    ;
 }
 
-hr {
-    color: black;
-}
-
-#mainTelaInicial {
-    width: 45px;
-    height: 809px;
+#fundoTelaInicial {
+    width: 365px;
+    height: 600px;
     border-radius: 10px;
     background: rgba(147, 144, 144, 0.39);
+    padding-top: 70px;
+    margin: 0 auto;
 }
 
 #emailTelaInicial,
 #senhaTelaInicial {
-    width: 434px;
-    height: 88px;
+    width: 300px;
+    height: 75px;
     flex-shrink: 0;
     border-radius: 8px;
     background: #FAFAFA;
     box-shadow: 0px 4px 4px 0px rgba(0, 0, 0, 0.25);
+    margin-left: 35px;
+    margin-top: 30px;
+}
+
+#senhaTelaInicial {
+    margin-bottom: 65px;
 }
 
 #emailDaTela,
 #senhaDaTela {
     color: #1E1E1E;
-    font-family: Inter, sans-serif;
+    font-family: 'Inter', sans-serif;
     font-size: 20px;
     font-style: normal;
     font-weight: 700;
     line-height: normal;
+    margin-bottom: -30px;
+    margin-left: 30px;
 }
 
 #botaoEntrar {
@@ -56,6 +64,9 @@ hr {
     font-size: 36px;
     font-style: normal;
     font-weight: 700px;
+    margin-left: 110px;
+    padding: 20px 25px 25px 30px;
+    text-decoration: none;
 }
 
 #entrarGoogle {
@@ -66,30 +77,31 @@ hr {
     font-style: normal;
     font-weight: 700;
     line-height: normal;
-    margin-top: 10px;
+    margin-top: 60px;
 }
 
 nav img {
     width: 65px;
     height: 65px;
     flex-shrink: 0;
-    margin-left: 180px;
+    margin-left: 160px;
 }
 
 #novaConta {
     display: flex;
-    align-items: center;
-    align-content: center;
-    width: 50%;
-    margin-top: -400px;
-    margin-left: 120px;
+    /* align-items: center;
+    align-content: center; */
+    width: 80%;
+    margin-top: -30px;
+    margin-left: 50px;
+
 }
 
-#criarConta {
-    color: rgba(102, 103, 100, 0.92);
+.criarConta {
+    color: rgba(102, 103, 100, 0.925);
     text-align: center;
     font-family: Inter, sans-serif;
-    font-size: 25px;
+    font-size: 20px;
     font-style: normal;
     font-weight: 700;
     line-height: normal;
@@ -105,4 +117,5 @@ nav img {
     font-weight: 700;
     line-height: normal;
     text-decoration-line: underline;
+    margin-left: 10px;
 }
